# swagger.yaml
openapi: 3.0.0
info:
  title: Cirkle API
  version: 1.0.0
  description: |
    ## Cirkle — Social Events Platform API

    Cirkle is an AI-native social discovery platform for people to find their tribe and attend events together.

    ### Authentication
    All endpoints except `GET /cities`, `GET /lifestyle-tags`, and `POST /auth/login` require a Bearer token.

    Include the token in every protected request:
    Authorization: Bearer <your_jwt_token>

    ### Price Format
    All prices are in **paise** (₹1 = 100 paise). Example: ₹500 = `50000`.

    ### Phone Format
    Indian mobile numbers only. Format: `+91XXXXXXXXXX` where the 10-digit number starts with 6, 7, 8, or 9.
servers:
  - url: http://localhost:3000
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /admin/admins:
    post:
      tags: [Admin — Admins]
      summary: Create an admin account
      description: |
        Creates a new admin (business_development or administrative). This is how administrative admins create BD team accounts.

        Requires the manage_admins capability — a business_development admin gets 403.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, displayName, role]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                displayName:
                  type: string
                role:
                  type: string
                  enum: [administrative, business_development]
      responses:
        '201':
          description: Admin created
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    $ref: '#/components/schemas/Admin'
        '400':
          description: Validation error (email format, password length, missing displayName, invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller lacks the manage_admins capability (business_development role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: You do not have permission to manage admin accounts
        '409':
          description: An admin with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [Admin — Admins]
      summary: List all admins
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List of admin accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  admins:
                    type: array
                    items:
                      $ref: '#/components/schemas/Admin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller lacks the manage_admins capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/admins/{id}:
    get:
      tags: [Admin — Admins]
      summary: Get a single admin account
      description: Same shape as the list rows. password_hash is never returned.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    $ref: '#/components/schemas/Admin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller lacks the manage_admins capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Admin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Admin — Admins]
      summary: Update an admin account
      description: |
        Partial update — only fields present in the body are changed. Commonly used to deactivate an admin (`isActive: false`), which invalidates their existing token immediately (authenticateAdmin re-checks is_active on every request, not just at token issuance).

        **Lockout protection.** Two rules make admin management unbrickable. Both return 409 with a machine-readable `error` so the client can say why:

        - `cannot_deactivate_self` — nobody may deactivate their own account, unconditionally, however many other admins exist. The failure mode is losing the session you are in.
        - `last_administrative_admin` — the population of ACTIVE administrative admins may never reach zero, whether by deactivating the last one or demoting them to business_development.

        Self-demotion is deliberately NOT blanket-blocked: it is governed by the second rule alone, so an administrative admin may hand over and step down to BD as long as somebody else remains administrative.

        The population count is taken under a row lock in the same transaction as the update, so two admins demoting each other simultaneously cannot both pass and leave zero.

        **Password reset.** Send `password` to reset it — alone, or alongside other fields. A body containing only `password` is valid and no longer returns "No fields to update". The value is bcrypt-hashed, never returned and never logged; the audit row records only that a reset occurred, not the value. A reset changes no counts, so the lockout rules never block it — resetting your own password is always allowed.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                role:
                  type: string
                  enum: [administrative, business_development]
                isActive:
                  type: boolean
                password:
                  type: string
                  minLength: 8
                  format: password
                  description: >-
                    New password. Hashed with bcrypt before storage, never
                    returned in any response and never written to logs. May be
                    the only field in the body.
      responses:
        '200':
          description: Admin updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    $ref: '#/components/schemas/Admin'
        '400':
          description: Validation error, or no fields provided to update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller lacks the manage_admins capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Admin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: |
            Blocked by lockout protection. Branch on `error`:

            - `cannot_deactivate_self` — you tried to deactivate your own account.
            - `last_administrative_admin` — the change would leave zero active administrative admins.
            - `concurrent_admin_change` — a competing admin change collided; retry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [cannot_deactivate_self, last_administrative_admin, concurrent_admin_change]
                  message:
                    type: string
                    description: Human-readable. Branch on `error`, not on this.
              examples:
                self:
                  value:
                    error: cannot_deactivate_self
                    message: You cannot deactivate your own account.
                last:
                  value:
                    error: last_administrative_admin
                    message: There must be at least one active administrative admin.
        '500':
          $ref: '#/components/responses/InternalError'
...
  /admin/invitations:
    get:
      tags: [Admin — Invitations]
      summary: List invitation activity across every organizer
      description: |
        The global view. The organizer dashboard will later show a scoped (own-events-only) version of this.

        **PII masking by role.** End-user `phone` and `email` are role-gated: `administrative` admins hold `view_pii` and see real values, `business_development` admins receive masked ones (`+91******210`, `j****@gmail.com`). Masking is server-side, so the real value never enters the response body.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: eventId
          schema:
            type: string
            format: uuid
          required: false
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, accepted, rejected]
          required: false
        - in: query
          name: userId
          schema:
            type: string
            format: uuid
          required: false
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Paginated list of invitations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginationEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AdminInvitation'
        '400':
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /organizer/events/{id}/invitations:
    get:
      tags: [Organizer]
      summary: List invitation requests for this event
      description: |
        Ownership-checked, same 404-not-403 rule as above. Filterable by status, defaulting to pending (the actionable set).

        Same profile-card treatment as attendees — no phone/email, ever.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, accepted, rejected]
            default: pending
          required: false
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Paginated list of invitation profile cards
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginationEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/OrganizerInvitation'
        '400':
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Event not found (also returned if it belongs to a different organizer)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /events/{id}/invitations:
    post:
      tags: [Events]
      summary: Request an invitation to an invite-only event
      description: |
        Creates a 'pending' invitation request for the logged-in user. Only valid for events with `eventType: invite_only`.

        A UNIQUE(user_id, event_id) constraint enforces one request per user per event — you cannot request twice while pending, re-request after rejection, or request again after acceptance. Attempting any of those returns 409 with the existing status.

        Approval/rejection happens outside this API (organizer tooling, out of scope). Poll `GET /events/{id}` to see `invitationStatus` change.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Event UUID
      responses:
        '201':
          description: Invitation request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: pending
        '400':
          description: Event is not invite-only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: This event does not require an invitation
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: |
            The event requires social handles the user hasn't supplied. Checked before the request row is created, so the organizer is guaranteed to see the required handles on every request card they decide on.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SocialHandlesRequired'
              example:
                error: social_handles_required
                missing: [instagram]
                message: This event requires your instagram handle.
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: A request already exists for this user and event
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  status:
                    type: string
                    enum: [pending, accepted, rejected]
              examples:
                alreadyPending:
                  value:
                    error: You have already requested an invitation
                    status: pending
                alreadyAccepted:
                  value:
                    error: Your invitation is already accepted
                    status: accepted
                alreadyRejected:
                  value:
                    error: Access to this event has been denied
                    status: rejected
        '500':
          $ref: '#/components/responses/InternalError'

# ─── COMPONENTS ───────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login. Valid for 7 days.
  parameters:
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        default: 50
        maximum: 100
        minimum: 1
      required: false
      description: Max rows to return. Default 50, clamped to 100 even if a larger value is requested.
    Offset:
      in: query
      name: offset
      schema:
        type: integer
        default: 0
        minimum: 0
      required: false
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Error message here
    Admin:
      type: object
      description: password_hash is never included in any response.
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          example: admin@cirkle.live
        displayName:
          type: string
          example: Founder Admin
        role:
          type: string
          enum: [administrative, business_development]
          description: Only capability difference is manage_admins — administrative has it, business_development doesn't. Otherwise identical access.
        isActive:
          type: boolean
          description: Kill switch. A deactivated admin's existing token stops working immediately, not just at expiry. Absent from the login response (implied true — login already checked it).
        createdAt:
          type: string
          format: date-time
...
    PaginationEnvelope:
      type: object
      description: Shared shape for every list endpoint. `data` is added per-endpoint via allOf.
      properties:
        total:
          type: integer
          description: Total matching rows, ignoring limit/offset — for "showing X-Y of Z" and page controls
          example: 3847
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 100
...
    AdminInvitation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, accepted, rejected]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        event:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        organizer:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        user:
          type: object
          description: PII — phone is unmasked for now.
          properties:
            id:
              type: string
              format: uuid
            firstName:
              type: string
              nullable: true
            age:
              type: integer
              nullable: true
            gender:
              type: string
              nullable: true
            phone:
              type: string
...
    OrganizerProfileCard:
      type: object
      description: |
        The base profile-card shape shared by OrganizerAttendee and OrganizerInvitation. Deliberately has NO phone and NO email properties — the backend's fetchAttendeeProfiles() never selects those columns in the first place, so there's nothing to leak even if a serializer bug existed. Organizers get profile cards only; Cirkle handles all attendee communication.

        Social handles (facebook/instagram/linkedin) ARE included: an event can require them before a user may buy or request an invitation, and the organizer needs to see them when deciding. That is a deliberate widening of the visible set — it does not extend to contact info, which remains absent by construction.
      properties:
        userId:
          type: string
          format: uuid
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        age:
          type: integer
          nullable: true
        gender:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        tagline:
          type: string
          nullable: true
        facebook:
          type: string
          nullable: true
          description: Bare handle, null if the attendee hasn't given one. Self-entered, not OAuth-verified.
        instagram:
          type: string
          nullable: true
          description: Bare handle, null if the attendee hasn't given one. Self-entered, not OAuth-verified.
        linkedin:
          type: string
          nullable: true
          description: Bare vanity slug, null if the attendee hasn't given one. Self-entered, not OAuth-verified.
        photos:
          type: array
          items:
            $ref: '#/components/schemas/CardPhoto'
        lifestyleTags:
          type: array
          items:
            $ref: '#/components/schemas/CardLifestyleTag'
    CardPhoto:
      type: object
      description: >-
        A photo as it appears on someone else's profile card (organizer
        attendee/invitation lists, vibe cards). Unlike the shared Photo schema
        this has NO id — those serializers never select it, since a viewer
        looking at another person's card has no reason to address an
        individual photo row.
      properties:
        url:
          type: string
          description: Presigned GET URL, ~1 hour TTL. The bucket is private — expires, don't cache long-term.
        position:
          type: integer
          minimum: 0
          maximum: 3
          description: Position 0 is the Main photo
    CardLifestyleTag:
      type: object
      description: >-
        A lifestyle tag as it appears on someone else's profile card. Unlike
        the shared LifestyleTag schema this has NO id.
      properties:
        label:
          type: string
          example: Clubbing
        category:
          type: string
          example: Going out
...
    OrganizerInvitation:
      description: An invitation request for the organizer's event. No phone/email — see OrganizerProfileCard.
      allOf:
        - type: object
          properties:
            invitationId:
              type: string
              format: uuid
            status:
              type: string
              enum: [pending, accepted, rejected]
            requestedAt:
              type: string
              format: date-time
        - $ref: '#/components/schemas/OrganizerProfileCard'
...
    SocialHandlesRequired:
      type: object
      description: >-
        The 403 returned by POST /payments/orders and POST /events/{id}/invitations
        when the event requires social handles the user hasn't supplied.

        Branch on `error` — it distinguishes this from every other 403 on those
        endpoints (an invite-only event needing an accepted invitation returns a
        different body). Use `missing` to decide which fields the handle dialog
        should show. Never parse `message`.
      properties:
        error:
          type: string
          enum: [social_handles_required]
        missing:
          type: array
          description: Exactly the required handles the user lacks, in a stable order.
          items:
            type: string
            enum: [facebook, instagram, linkedin]
          example: [instagram, linkedin]
        message:
          type: string
          description: Human-readable, for logging or a fallback. Not a stable contract.
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: No token provided
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error
...